refactor: change browser to utils
