+++ lua/browse/mdn.lua
@@ -1,7 +1,7 @@
-local browser = require("browse.browser")
+local utils = require("browse.utils")
 
 local M = {}
 
-M.search = browser.generic_search_for("https://developer.mozilla.org/en-US/search?q=%s")
+M.search = utils.generic_search_for("https://developer.mozilla.org/en-US/search?q=%s")
 
 return M
